Create rectangle calculator.py
this script is a basic script made to show the area of a rectangle. First defining a calculate area function to be able to use in the following calculate function. First creating a while true loop for valid inputs, once valid inputs are received then is will create 'area_sq_cm' which contains the area value which can then be used in the following text